[SwiftLint] Breaks out Moya.swift into multiple files.
