+++ Source/Moya+Defaults.swift
@@ -1,0 +1,24 @@
+import Alamofire
+
+public extension MoyaProvider {
+
+    // These functions are default mappings to MoyaProvider's properties: endpoints, requests, manager, etc.
+
+    public final class func DefaultEndpointMapping(target: Target) -> Endpoint<Target> {
+        let url = target.baseURL.URLByAppendingPathComponent(target.path).absoluteString
+        return Endpoint(URL: url, sampleResponseClosure: {.NetworkResponse(200, target.sampleData)}, method: target.method, parameters: target.parameters)
+    }
+
+    public final class func DefaultRequestMapping(endpoint: Endpoint<Target>, closure: RequestResultClosure) {
+        return closure(.Success(endpoint.urlRequest))
+    }
+
+    public final class func DefaultAlamofireManager() -> Manager {
+        let configuration = NSURLSessionConfiguration.defaultSessionConfiguration()
+        configuration.HTTPAdditionalHeaders = Manager.defaultHTTPHeaders
+
+        let manager = Manager(configuration: configuration)
+        manager.startRequestsImmediately = false
+        return manager
+    }
+}
